+++ Autopilot/AttitudeManager/Src/attitudeStateClasses.cpp
@@ -59,28 +59,42 @@ void PIDloopMode::execute(attitudeManager* attitudeMgr)
     PMError_t pmError = PM_GetCommands(&pathManagerOutput);
 
     //executes PID's to acheive desired roll, pitch angle
-    _PidOutput.rollPercent = _rollPid.execute(PMInstructions->roll, SFOutput->IMUroll, SFOutput->IMUrollrate);
-    _PidOutput.pitchPercent = _pitchPid.execute(PMInstructions->pitch, SFOutput->IMUpitch, SFOutput->IMUpitchrate);
-    _PidOutput.rudderPercent = pathManagerOutput.rudderPercent;
-    _PidOutput.throttlePercent = pathManagerOutput.throttlePercent;
-
     //if manual control is needed, use loaded in percents instead!
     if(pathManagerOutput.passbyData.pitchPassby)
     {
         _PidOutput.pitchPercent = pathManagerOutput.passbyData.pitchPercent;
-    } 
+    }
+    else
+    {
+        _PidOutput.pitchPercent = _pitchPid.execute(PMInstructions->pitch, SFOutput->IMUpitch, SFOutput->IMUpitchrate);
+    }
+
     if(pathManagerOutput.passbyData.rollPassby)
     {
         _PidOutput.rollPercent = pathManagerOutput.passbyData.rollPercent;
-    } 
+    }
+    else
+    {
+        _PidOutput.rollPercent = _rollPid.execute(PMInstructions->roll, SFOutput->IMUroll, SFOutput->IMUrollrate);
+    }
+
     if(pathManagerOutput.passbyData.rudderPassby)
     {
         _PidOutput.rudderPercent = pathManagerOutput.passbyData.rudderPercent;
-    } 
+    }
+    else
+    {
+        _PidOutput.rudderPercent = pathManagerOutput.rudderPercent;
+    }
+
     if(pathManagerOutput.passbyData.throttlePassby)
     {
         _PidOutput.throttlePercent = pathManagerOutput.passbyData.throttlePercent;
     } 
+    else
+    {
+        _PidOutput.throttlePercent = pathManagerOutput.throttlePercent;
+    }
 
     if (pmError.errorCode == 0) 
     {
