+++ Autopilot/PathManager/Inc/pathDatatypes.hpp
@@ -52,4 +52,23 @@ struct Telemetry_PIGO_t {
 
 };
 
+/*
+* Structure stores manual control information
+* If a certain variable needs to be controlled, set the percent to the desired value and its respective boolean to true
+*/
+struct _PassbyControl{
+    double rollPercent;
+    bool rollPassby;
+
+    double rudderPercent;
+    bool rudderPassby;
+
+    double pitchPercent;
+    bool pitchPassby;
+
+    double throttlePercent;
+    bool throttlePassby;
+};
+
 #endif
+
